+++ proj.android/ChatServices/src/com/elex/chatservice/model/ChannelListItem.java
@@ -1,17 +1,16 @@
 package com.elex.chatservice.model;
 
-
 public class ChannelListItem
 {
-	public transient boolean checked = false;
-	public transient int			unreadCount		= 0;
-	
+	public transient boolean	checked		= false;
+	public transient int		unreadCount	= 0;
+
 	public boolean isUnread()
 	{
 		return false;
 	}
-	
-	public int getChannelTime()
+
+	public long getChannelTime()
 	{
 		return 0;
 	}
